DRTII-388 cruching file upload for LGW with arrivalsActor
