+++ server/src/main/scala/drt/server/feeds/common/ArrivalFeed.scala
@@ -1,0 +1,30 @@
+package drt.server.feeds.common
+
+import akka.actor.ActorRef
+import akka.pattern.ask
+import akka.util.Timeout
+import drt.shared.FlightsApi.Flights
+import org.slf4j.{Logger, LoggerFactory}
+import server.feeds.{ArrivalsFeedFailure, ArrivalsFeedResponse, ArrivalsFeedSuccess, GetFeedImportArrivals}
+import services.SDate
+
+import scala.concurrent.ExecutionContext.Implicits.global
+import scala.concurrent.Future
+
+case class ArrivalFeed(arrivalsActor: ActorRef)(implicit timeout: Timeout) {
+  val log: Logger = LoggerFactory.getLogger(getClass)
+
+  def requestFeed(portCode: String): Future[ArrivalsFeedResponse] =
+    arrivalsActor.ask(GetFeedImportArrivals)
+      .map {
+        case Some(Flights(arrivals)) =>
+          log.info(s"Got ${arrivals.length} $portCode port arrivals")
+          ArrivalsFeedSuccess(Flights(arrivals), SDate.now())
+        case x =>
+          log.info(s"Got no $portCode port arrivals: $x")
+          ArrivalsFeedSuccess(Flights(Seq()), SDate.now())
+      }
+      .recoverWith {
+        case e => Future(ArrivalsFeedFailure(e.getMessage, SDate.now()))
+      }
+}
