+++ server/src/test/scala/drt/server/feeds/lhr/LHRForecastFeedSpec.scala
@@ -2,6 +2,7 @@ package drt.server.feeds.lhr
 
 import akka.actor.{Actor, Props}
 import akka.util.Timeout
+import drt.server.feeds.common.ArrivalFeed
 import server.feeds.ArrivalsFeedFailure
 import services.crunch.CrunchTestLike
 
@@ -19,10 +20,10 @@ class SlowResponseActor extends Actor {
 class LHRForecastFeedSpec extends CrunchTestLike {
   "Given an LHR forecast feed with a mock actor that takes longer than the timeout to respond" >> {
     val mockActor = system.actorOf(Props(new SlowResponseActor()))
-    val lhrFeed = LHRForecastFeed(mockActor)(new Timeout(100 milliseconds))
+    val lhrFeed = ArrivalFeed(mockActor)(new Timeout(100 milliseconds))
 
     "When I request the feed" >> {
-      val result = Await.result(lhrFeed.requestFeed, 2 seconds)
+      val result = Await.result(lhrFeed.requestFeed("LHR"), 2 seconds)
 
       "I should get a ArrivalsFeedFailure rather than an uncaught exception" >> {
         result.getClass === classOf[ArrivalsFeedFailure]
